Develop

+++ src/Http/Controllers/DiscussionController.php
@@ -53,6 +53,8 @@ class DiscussionController extends Controller
      */
     public function store(StoreDiscussionRequest $request, Group $group)
     {
+        $this->authorize('create', Discussion::class);
+
         $discussion = $this->discussionService->make($request, $group);
 
         return redirect()->route('firefly.discussion.show', [$discussion->id, $discussion->slug]);
@@ -92,6 +94,8 @@ class DiscussionController extends Controller
      */
     public function update(UpdateDiscussionRequest $request, Discussion $discussion)
     {
+        $this->authorize('update', $discussion);
+
         $discussion = $this->discussionService->update($request, $discussion);
 
         return redirect()->route('firefly.discussion.show', [$discussion->id, $discussion->slug]);
@@ -105,6 +109,8 @@ class DiscussionController extends Controller
      */
     public function delete(Request $request, Discussion $discussion)
     {
+        $this->authorize('delete', $discussion);
+
         $this->discussionService->delete($discussion);
         
         return redirect()->route('firefly.index');
@@ -119,6 +125,8 @@ class DiscussionController extends Controller
      */
     public function lock(Request $request, Discussion $discussion)
     {
+        $this->authorize('lock', $discussion);
+
         $this->discussionService->updateState($discussion, 'lock');
 
         return redirect()->route('firefly.discussion.show', [$discussion->id, $discussion->slug]);
@@ -133,6 +141,8 @@ class DiscussionController extends Controller
      */
     public function unlock(Request $request, Discussion $discussion)
     {
+        $this->authorize('unlock', $discussion);
+
         $this->discussionService->updateState($discussion, 'unlock');
 
         return redirect()->route('firefly.discussion.show', [$discussion->id, $discussion->slug]);
@@ -147,6 +157,8 @@ class DiscussionController extends Controller
      */
     public function pin(Request $request, Discussion $discussion)
     {
+        $this->authorize('pin', $discussion);
+
         $this->discussionService->updateState($discussion, 'pin');
 
         return redirect()->route('firefly.discussion.show', [$discussion->id, $discussion->slug]);
@@ -161,6 +173,8 @@ class DiscussionController extends Controller
      */
     public function unpin(Request $request, Discussion $discussion)
     {
+        $this->authorize('unpin', $discussion);
+
         $this->discussionService->updateState($discussion, 'unpin');
 
         return redirect()->route('firefly.discussion.show', [$discussion->id, $discussion->slug]);
